# refactor: Optimize DecryptNotes worker job