## Summary

This was a PR I started a few weeks ago, but put on hold after the blstrs revert and stuff. I was using the DecryptNotes job as a test-bed for the rust worker pool investigation, and while working on it, I discovered a few ways to make the DecryptNotes job a bit cleaner and more optimized.

Baseline performance was mostly unchanged (i.e. when the job has 0 or 1 decryptable notes), but this will improve performance when there are multiple successfully decryptable notes, so scenarios where an account received more than 1 note in a batch, like splitting notes or a popular service receiving multiple transfers in the same batch.

The main changes:
- Removed the null values in the response - we were skipping these everywhere it was used, so it seemed pointless to keep them, it's just using more memory and creating larger payloads to de/serialize between the worker and pool.
- We were sending the keys with every note in the payload, despite always only using the 1 key. By separating the main payload data from the actual encrypted notes, we dramatically decrease the size of the payload. 
  - This also makes it easier in the future to support multiple keys, as another future improvement. I hope that would improve sync/rescan/decrypt speeds for nodes with multiple accounts. Right now, we kick off the set of notes for each key, but by sending multiple keys, we would be able to skip decrypting notes for other accounts if one already successfully decrypted it, since a note can only have 1 owner.
- Only instantiate the key objects once, and as needed. Previously, we were initializing the key objects with every note

### TODOs
- [x] Update the perf test to the new format
- [x] De-dupe some of the logic with the new functions
- [x] Outline the perf changes again with the portable blstrs changes

## Performance changes

Ran on a c6i.large (2vCPU, 4gb RAM).

To summarize: CPU performance increases a small amount when multiple notes can be decrypted. Otherwise, there is minor-to-no change when no notes can be decrypted (which is the vast majority of cases). Memory usage is a bit lower as well. The main goal here was no measurable CPU regressions in the main case (no decryptable notes) and an overall improvement in memory usage.

### Existing

```
[DecryptNotes: notes: 20, canDecrypt: true, decryptForSpender: true]
Iterations: 100
Time: min: 40.1665ms, avg: 40.8651ms, median: 40.4997ms, max 44.5321ms
Heap: min: -13055912 B, avg: -89005 B, median: 39.08 KiB, max 61.53 KiB
Rss: min: 0 B, avg: 89.00 KiB, median: 0 B, max 792.00 KiB
Mem: min: -12785576 B, avg: 2.08 KiB, median: 39.08 KiB, max 831.18 KiB

[DecryptNotes: notes: 20, canDecrypt: true, decryptForSpender: false]
Iterations: 100
Time: min: 40.1389ms, avg: 40.6928ms, median: 40.3954ms, max 44.489ms
Heap: min: 232.24 KiB, avg: 40.82 KiB, median: 38.10 KiB, max 57.25 KiB
Rss: min: 0 B, avg: 31.96 KiB, median: 0 B, max 448.00 KiB
Mem: min: 182.96 KiB, avg: 72.78 KiB, median: 37.96 KiB, max 57.25 KiB

[DecryptNotes: notes: 20, canDecrypt: false, decryptForSpender: true]
Iterations: 100
Time: min: 11.5717ms, avg: 11.7239ms, median: 11.6245ms, max 14.2146ms
Heap: min: 17.39 KiB, avg: 20.59 KiB, median: 18.25 KiB, max 35.53 KiB
Rss: min: 0 B, avg: 14.12 KiB, median: 0 B, max 264.00 KiB
Mem: min: 158.25 KiB, avg: 34.71 KiB, median: 18.25 KiB, max 518.11 KiB

[DecryptNotes: notes: 20, canDecrypt: false, decryptForSpender: false]
Iterations: 100
Time: min: 7.1689ms, avg: 7.2937ms, median: 7.2301ms, max 9.8428ms
Heap: min: 17.39 KiB, avg: 17.65 KiB, median: 17.39 KiB, max 35.51 KiB
Rss: min: 0 B, avg: 14.64 KiB, median: 0 B, max 264.00 KiB
Mem: min: 17.39 KiB, avg: 32.29 KiB, median: 17.39 KiB, max 35.51 KiB

[DecryptNotes: notes: 100, canDecrypt: true, decryptForSpender: true]
Iterations: 100
Time: min: 203.9769ms, avg: 205.4167ms, median: 204.5078ms, max 226.4551ms
Heap: min: -13133776 B, avg: 3.08 KiB, median: 137.72 KiB, max 158.20 KiB
Rss: min: -17952768 B, avg: -400343 B, median: 0 B, max 788.00 KiB
Mem: min: -1567960 B, avg: -397182 B, median: 155.86 KiB, max 925.72 KiB

[DecryptNotes: notes: 100, canDecrypt: true, decryptForSpender: false]
Iterations: 100
Time: min: 204.0819ms, avg: 213.6366ms, median: 205.1782ms, max 459.923ms
Heap: min: -1906584 B, avg: -414824 B, median: 137.61 KiB, max 269.01 KiB
Rss: min: -1024000 B, avg: -501924 B, median: 190.00 KiB, max 528.00 KiB
Mem: min: -1031560 B, avg: -916748 B, median: 348.46 KiB, max 668.17 KiB

[DecryptNotes: notes: 100, canDecrypt: false, decryptForSpender: true]
Iterations: 100
Time: min: 57.2586ms, avg: 57.6199ms, median: 57.4283ms, max 60.366ms
Heap: min: -832176 B, avg: -4092 B, median: 42.64 KiB, max 64.42 KiB
Rss: min: -1048576 B, avg: 98.76 KiB, median: 0 B, max 792.00 KiB
Mem: min: -1473344 B, avg: 94.76 KiB, median: 42.43 KiB, max 834.64 KiB

[DecryptNotes: notes: 100, canDecrypt: false, decryptForSpender: false]
Iterations: 100
Time: min: 35.3373ms, avg: 35.6544ms, median: 35.4139ms, max 40.3348ms
Heap: min: -791504 B, avg: -3428 B, median: 42.64 KiB, max 60.81 KiB
Rss: min: -1048576 B, avg: 32.32 KiB, median: 0 B, max 396.00 KiB
Mem: min: -1631184 B, avg: 28.97 KiB, median: 42.46 KiB, max 60.57 KiB
```

### New

```
    [DecryptNotes: notes: 20, canDecrypt: true, decryptForSpender: true]
    Iterations: 100
    Time: min: 33.9509ms, avg: 34.5123ms, median: 34.322ms, max 37.9586ms
    Heap: min: -13632344 B, avg: -101520 B, median: 33.51 KiB, max 51.73 KiB
    Rss: min: 0 B, avg: 73.44 KiB, median: 0 B, max 88.00 KiB
    Mem: min: -13439832 B, avg: -26318 B, median: 33.51 KiB, max 51.73 KiB

    [DecryptNotes: notes: 20, canDecrypt: true, decryptForSpender: false]
    Iterations: 100
    Time: min: 33.8926ms, avg: 34.395ms, median: 34.2046ms, max 37.7969ms
    Heap: min: 32.32 KiB, avg: 33.64 KiB, median: 33.18 KiB, max 51.66 KiB
    Rss: min: 0 B, avg: 2.36 KiB, median: 0 B, max 64.00 KiB
    Mem: min: 99.27 KiB, avg: 36.00 KiB, median: 33.18 KiB, max 52.40 KiB

    [DecryptNotes: notes: 20, canDecrypt: false, decryptForSpender: true]
    Iterations: 100
    Time: min: 11.5488ms, avg: 11.68ms, median: 11.636ms, max 12.5578ms
    Heap: min: 11.17 KiB, avg: 12.51 KiB, median: 12.07 KiB, max 31.19 KiB
    Rss: min: 0 B, avg: 3.16 KiB, median: 0 B, max 8.00 KiB
    Mem: min: 11.17 KiB, avg: 15.67 KiB, median: 12.13 KiB, max 31.19 KiB

    [DecryptNotes: notes: 20, canDecrypt: false, decryptForSpender: false]
    Iterations: 100
    Time: min: 7.1516ms, avg: 7.2395ms, median: 7.2018ms, max 8.0417ms
    Heap: min: 11.17 KiB, avg: 11.36 KiB, median: 11.17 KiB, max 29.25 KiB
    Rss: min: 0 B, avg: 0 B, median: 0 B, max 0 B
    Mem: min: 11.17 KiB, avg: 11.36 KiB, median: 11.17 KiB, max 29.25 KiB

    [DecryptNotes: notes: 100, canDecrypt: true, decryptForSpender: true]
    Iterations: 100
    Time: min: 170.3022ms, avg: 172.0129ms, median: 171.3873ms, max 192.9646ms
    Heap: min: -13245968 B, avg: -19007 B, median: 108.63 KiB, max 127.73 KiB
    Rss: min: -1036288 B, avg: -117514 B, median: 0 B, max 792.00 KiB
    Mem: min: -13245968 B, avg: -136521 B, median: 108.68 KiB, max 915.73 KiB

    [DecryptNotes: notes: 100, canDecrypt: true, decryptForSpender: false]
    Iterations: 100
    Time: min: 170.4706ms, avg: 175.2092ms, median: 171.6689ms, max 424.2823ms
    Heap: min: -1452048 B, avg: -9454 B, median: 109.42 KiB, max 234.10 KiB
    Rss: min: -1019904 B, avg: -223027 B, median: 0 B, max 76.00 KiB
    Mem: min: -1082128 B, avg: -232481 B, median: 114.10 KiB, max 718.57 KiB

    [DecryptNotes: notes: 100, canDecrypt: false, decryptForSpender: true]
    Iterations: 100
    Time: min: 56.8091ms, avg: 57.2968ms, median: 56.9924ms, max 63.0896ms
    Heap: min: -930000 B, avg: -6389 B, median: 11.17 KiB, max 29.29 KiB
    Rss: min: -1015808 B, avg: 137.12 KiB, median: 0 B, max 528.00 KiB
    Mem: min: -1701832 B, avg: 130.88 KiB, median: 12.14 KiB, max 539.17 KiB

    [DecryptNotes: notes: 100, canDecrypt: false, decryptForSpender: false]
    Iterations: 100
    Time: min: 34.9813ms, avg: 35.2289ms, median: 35.0502ms, max 37.6621ms
    Heap: min: -773496 B, avg: -3802 B, median: 11.17 KiB, max 29.29 KiB
    Rss: min: -741376 B, avg: -3031 B, median: 0 B, max 264.00 KiB
    Mem: min: -1514872 B, avg: -6833 B, median: 11.17 KiB, max 29.29 KiB
```

## Testing Plan

## Documentation

Does this change require any updates to the Iron Fish Docs (ex. [the RPC API
Reference](https://ironfish.network/docs/onboarding/rpc/chain))? If yes, link a
related documentation pull request for the website.

```
[ ] Yes
```

## Breaking Change

Is this a breaking change? If yes, add notes below on why this is breaking and
what additional work is required, if any.

```
[ ] Yes
```